+++ tests/helpers/templates/sdist/standard_default_vcs_mercurial_exclusion_files.py
@@ -36,6 +36,7 @@ syntax: glob
 Metadata-Version: {DEFAULT_METADATA_VERSION}
 Name: {kwargs['project_name']}
 Version: 0.0.1
+License-File: LICENSE.txt
 """,
         )
     )