+++ tests/helpers/templates/wheel/standard_editable_exact.py
@@ -47,6 +47,7 @@ Tag: py3-none-any
 Metadata-Version: {DEFAULT_METADATA_VERSION}
 Name: {kwargs['project_name']}
 Version: 0.0.1
+License-File: LICENSE.txt
 Requires-Dist: editables~=0.3
 """,
         )