+++ tests/helpers/templates/wheel/standard_editable_pth_force_include.py
@@ -39,6 +39,7 @@ Tag: py3-none-any
 Metadata-Version: {DEFAULT_METADATA_VERSION}
 Name: {kwargs['project_name']}
 Version: 0.0.1
+License-File: LICENSE.txt
 """,
         )
     )