+++ docs/history.md
@@ -172,6 +172,7 @@ This is the first stable release of Hatch v1, a complete rewrite. Enjoy!
 
 - Add `metadata` command to view PEP 621 project metadata
 - Improve error messages for SPDX license errors
+- Retroactively support `License-File` for core metadata starting at version 2.1
 
 ***Fixed:***